fix delete category

+++ src/main/java/com/example/hadistore/service/category/CategoryService.java
@@ -2,12 +2,14 @@ package com.example.hadistore.service.category;
 
 import com.example.hadistore.dtos.CategoryDTO;
 import com.example.hadistore.entity.Category;
+import com.example.hadistore.exceptions.DataNotFoundException;
 import com.example.hadistore.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -35,12 +37,16 @@ public class CategoryService implements ICategoryService{
     @Override
     @Transactional
     public void deleteCategory(Integer id) {
+        Optional<Category> existCategory = categoryRepository.findById(id);
+        if (existCategory.isEmpty()){
+            throw new DataNotFoundException("Category not found");
+        }
         categoryRepository.deleteById(id);
     }
 
     @Override
     public Category getCategoryById(Integer id) {
-        return categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Category not found"));
+        return categoryRepository.findById(id).orElseThrow(() -> new DataNotFoundException("Category not found"));
     }
 
     @Override
